refactor(cli): extract session attachments
