+++ packages/agent-cli/src/Agent/CLI/Runtime.hs
@@ -106,8 +106,7 @@ import Agent.CLI.Btw
     )
 import Agent.CLI.CancelWatch (withEscCancel, withStdinPaused)
 import Agent.CLI.Clipboard
-    ( appendUniqueImageAttachments
-    , formatImageSize
+    ( formatImageSize
     , loadImagesFromPastedText
     , nonEmptyClipboardImages
     , readClipboardImagesForPaste
@@ -150,12 +149,6 @@ import Agent.CLI.Error
     ( formatApiErrorAt
     , formatApiErrorInlineAt
     , formatApiErrorRetryCountdownParts
-    )
-import Agent.CLI.ImagePreview
-    ( detectImagePreviewProtocol
-    , previewColumnsFor
-    , previewRowsFor
-    , renderImagePreview
     )
 import Agent.CLI.Input
     ( ReplLine(..)
@@ -337,6 +330,11 @@ import Agent.CLI.Session.Choices
     , effortChoice
     , modelChoice
     , showAccountUsage
+    )
+import Agent.CLI.Session.Attachments
+    ( putImagePreview
+    , queueAttachedImages
+    , queueClipboardImages
     )
 import Agent.CLI.Session.History
     ( detectGitBranch
@@ -6033,88 +6031,6 @@ setNativeProgress handle active = do
                 | otherwise = osc9ProgressRemove
         Text.hPutStr handle (wrapOscForTmux inTmux sequence_)
         hFlush handle
-
--- | Queue clipboard / Finder-paste images and optionally draw an in-terminal
--- thumbnail. The caller reports the returned message through the active UI.
-queueAttachedImages
-    :: IORef [ImageAttachment]
-    -> IORef Int
-    -> Bool
-    -> Bool
-    -> [ImageAttachment]
-    -> IO Text
-queueAttachedImages attachmentsRef previewIdRef color showPreview images = do
-    (added, pendingCount) <- atomicModifyIORef' attachmentsRef \existing ->
-        let (pending, unique) =
-                appendUniqueImageAttachments existing images
-        in (pending, (unique, length pending))
-    let sizes =
-            Text.intercalate ", "
-                [ img.imageMime <> " (" <> formatImageSize (BS.length img.imageBytes) <> ")"
-                | img <- added
-                ]
-        duplicateCount = length images - length added
-        queued = Text.pack (show pendingCount)
-    when (showPreview && not (null added)) $
-        putImagePreview previewIdRef color added
-    pure $ case (added, duplicateCount) of
-        ([], _) ->
-            "image already attached — not added again ("
-                <> queued
-                <> " queued)"
-        (_, 0) ->
-            "attached "
-                <> sizes
-                <> " — send with next message ("
-                <> queued
-                <> " queued)"
-        _ ->
-            "attached "
-                <> sizes
-                <> "; skipped "
-                <> Text.pack (show duplicateCount)
-                <> " duplicate image"
-                <> (if duplicateCount == 1 then "" else "s")
-                <> " ("
-                <> queued
-                <> " queued)"
-
-queueClipboardImages
-    :: IORef [ImageAttachment]
-    -> IORef Int
-    -> Bool
-    -> Bool
-    -> IO (Either Text Text)
-queueClipboardImages
-    attachmentsRef
-    previewIdRef
-    color
-    showPreview = do
-    imagesResult <- readClipboardImagesForPaste
-    case imagesResult of
-        Right images@(_:_) ->
-            Right <$> queueAttachedImages
-                attachmentsRef previewIdRef color showPreview images
-        Right [] ->
-            pure (Left "no image found on the clipboard")
-        Left err -> pure (Left err)
-
-putImagePreview :: IORef Int -> Bool -> [ImageAttachment] -> IO ()
-putImagePreview previewIdRef color images = do
-    protocol <- detectImagePreviewProtocol stdout
-    inTmux <- isJust <$> lookupEnv "TMUX"
-    size <- getTerminalSize
-    let (termRows, termCols) = fromMaybe (24, 80) size
-        columns = previewColumnsFor termCols
-        rows = previewRowsFor termRows
-    startId <- atomicModifyIORef' previewIdRef \n ->
-        (n + max 1 (length images), n)
-    case renderImagePreview protocol inTmux (roleMuted color) columns rows startId images of
-        Nothing -> pure ()
-        Just block -> do
-            -- Graphics sequences must not go through the Solarized wash.
-            Text.putStrLn block
-            hFlush stdout
 
 putTrailingNewline :: IORef Bool -> IO ()
 putTrailingNewline printed = do
